requirements.txt
jupyter
matplotlib
ipython-sql
version_information
numpy
